requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.26.0
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Delorean==1.0.0
dotenv-python==0.0.1
et-xmlfile==1.1.0
frozenlist==1.3.0
gino==1.0.1
h11==0.13.0
humanize==4.2.3
idna==3.3
multidict==6.0.2
numpy==1.23.1
openpyxl==3.0.10
openpyxl-image-loader==1.0.5
outcome==1.2.0
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
wsproto==1.1.0
xlrd==1.2.0
yarl==1.7.2
